new package persistence class UserRepository moved to package persistence class User moved to package server
