+++ src/persistence/UserRepository.java
@@ -1,4 +1,6 @@
-package server.authorization.users;
+package persistence;
+
+import server.User;
 
 import java.sql.*;
 import java.util.ArrayList;
